Moving validationSession logic into Session Provider, review routes name
