+++ public/robots.txt
@@ -1,25 +1,4 @@
 User-agent: *
-Disallow: /
-Disallow: /access
-Disallow: /accessL3
-Disallow: /logoutInit
-Disallow: /session
-Disallow: /profile
-Disallow: /profileBlock
-Disallow: /profileRestore
-Disallow: /profileRestoreL3
-Disallow: /restoreCode
-Disallow: /restoreThankYou
-Disallow: /profileBlockSuccess
-Disallow: /profileAccessBlock
-Disallow: /validateSession
-Disallow: /thankyou
-Disallow: /logoutKoAuth
-Disallow: /logoutKo
-Disallow: /ko
-Disallow: /logoutNoSession/l1
-Disallow: /logoutNoSession/l2
-Disallow: /error
 
 Disallow: /
 Disallow: /accedi
@@ -39,6 +18,8 @@ Disallow: /ripristino-accesso/inserisci-codice
 Disallow: /ripristino-accesso/operazione-completata
 Disallow: /ripristino-accesso/errore
 Disallow: /ripristino-accesso/accedi-livello-3
+Disallow: /expiredMagicLink
+
