+++ src/middleware.ts
@@ -1,5 +1,6 @@
 import createIntlMiddleware from 'next-intl/middleware';
-import { NextRequest } from 'next/server';
+import { NextRequest, NextResponse } from 'next/server';
+import { EXISTING_ROUTES, ROUTES } from './app/[locale]/_utils/routes';
 
 export default async function middleware(request: NextRequest) {
   const defaultLocale = request.headers.get('x-default-locale') || 'en';
@@ -12,6 +13,10 @@ export default async function middleware(request: NextRequest) {
   const response = handleI18nRouting(request);
 
   response.headers.set('x-default-locale', defaultLocale);
+
+  if (!EXISTING_ROUTES.includes(request.nextUrl.pathname)) {
+    return NextResponse.redirect(new URL(ROUTES.LOGIN, request.url));
+  }
 
   return response;
 }
